Struct rocket_cors::Guard

pub struct Guard<'r> { /* fields omitted */ }

A request guard to check CORS headers before a route is run. Will not execute the route if checks fail.

See the documentation at the crate root for usage information.

You should not wrap this in an Option or Result because the guard will let non-CORS requests through and will take over error handling in case of errors. In essence, this is just a wrapper around Response with a 'r borrowed lifetime so users don't have to keep specifying the lifetimes in their routes

Implementations

impl<'r> Guard<'r>

pub fn responder<R: Responder<'r>>(self, responder: R) -> Responder<'r, R>

Consumes the Guard and return a Responder that wraps a provided rocket:response::Responder with CORS headers

pub fn response(&self, base: Response<'r>) -> Response<'r>

Merge a rocket::Response with this CORS Guard. This is usually used in the final step of a route to return a value for the route.

This will overwrite any existing CORS headers