Enum rocket_cors::Error

pub enum Error {
    MissingOrigin,
    BadOrigin(ParseError),
    OpaqueAllowedOrigin(Vec<String>),
    MissingRequestMethod,
    BadRequestMethod,
    MissingRequestHeaders,
    OriginNotAllowed(String),
    MethodNotAllowed(String),
    RegexError(Error),
    HeadersNotAllowed,
    CredentialsWithWildcardOrigin,
    MissingCorsInRocketState,
    MissingInjectedHeader,
}

Errors during operations

This enum implements rocket::response::Responder which will return an appropriate status code while printing out the error in the console. Because these errors are usually the result of an error while trying to respond to a CORS request, CORS headers cannot be added to the response and your applications requesting CORS will not be able to see the status code.

Variants

MissingOrigin

The HTTP request header Origin is required but was not provided

BadOrigin(ParseError)

The HTTP request header Origin could not be parsed correctly.

OpaqueAllowedOrigin(Vec<String>)

The configured Allowed Origins are Opaque origins. Use a Regex instead.

MissingRequestMethod

The request header Access-Control-Request-Method is required but is missing

BadRequestMethod

The request header Access-Control-Request-Method has an invalid value

MissingRequestHeaders

The request header Access-Control-Request-Headers is required but is missing.

OriginNotAllowed(String)

Origin is not allowed to make this request

MethodNotAllowed(String)

Requested method is not allowed

RegexError(Error)

A regular expression compilation error

HeadersNotAllowed

One or more headers requested are not allowed

CredentialsWithWildcardOrigin

Credentials are allowed, but the Origin is set to "*". This is not allowed by W3C

This is a misconfiguration. Check the docuemntation for Cors.

MissingCorsInRocketState

A CORS Request Guard was used, but no CORS Options was available in Rocket's state

This is a misconfiguration. Use Rocket::manage to add a CORS options to managed state.

MissingInjectedHeader

The on_response handler of Fairing could not find the injected header from the Request. Either some other fairing has removed it, or this is a bug.