Struct rocket_cors::headers::AccessControlRequestHeaders

pub struct AccessControlRequestHeaders(pub HeaderFieldNamesSet);

The Access-Control-Request-Headers request header

You can use this as a rocket Request Guard to ensure that the header is passed in correctly.